updated and moved delete_rating and updated and moved tests so that they're under the proper app
