+++ accounts/templates/accounts/profile.html
@@ -42,13 +42,12 @@
 
 						<div id="collapse-{{rating.pk}}" class="collapse" aria-labelledby="heading-{{rating.pk}}" data-parent="#accordionExample">
 							<div class="card-body">
-                                <li>Restroom: {{ rating.restroom_id_id }}</li>
 								<li>Rating: {{rating.rating}}</li>
 								<li>Headline: {{rating.headline}}</li>
 								<li>Comment: {{rating.comment}}</li>
 {#								<a href="{% url 'accounts:edit_rating' rating.id %}" >Edit</a><br>#}
 								<a href="{% url 'naturescall:rate_restroom' rating.restroom_id_id %}" >Edit</a><br>
-								<a href="{% url 'accounts:delete_rating' rating.id %}" >Delete</a>
+								<a href="{% url 'naturescall:delete_rating' rating.restroom_id_id %}" >Delete</a>
 							</div>
 						</div>
 					</div>
