+++ naturescall/tests.py
@@ -137,13 +137,6 @@ class ViewTests(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(str(response.content).count("Add Restroom"), 19)
 
-    def test_access_signup(self):
-        """
-        A get request to the signup page should yield a valid response
-        """
-        response = self.client.get(reverse("accounts:signup"))
-        self.assertEqual(response.status_code, 200)
-
     def test_get_request_add_restroom_not_logged_in(self):
         """
         A get request to the add_restroom page should yield a
@@ -174,50 +167,6 @@ class ViewTests(TestCase):
         yelp_id = "E6h-sMLmF86cuituw5zYxwXXXXXX"
         response = self.client.get(reverse("naturescall:add_restroom", args=(yelp_id,)))
         self.assertEqual(response.status_code, 404)
-
-    def test_account_creation_valid_form(self):
-        """
-        A valid form should yield a redirect upon submission and
-        add a user to the database
-        """
-        response = self.client.post(
-            reverse("accounts:signup"),
-            data={
-                "username": "test_user",
-                "email": "test_user@email.com",
-                "first_name": "test",
-                "last_name": "user",
-                "password1": "BDbdKDwpSt",
-                "password2": "BDbdKDwpSt",
-            },
-        )
-        all_users = User.objects.filter(id=1)
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(len(all_users), 1)
-
-    def test_account_creation_invalid_form(self):
-        """
-        An invalid form should yield an error upon submission
-        """
-        response = self.client.post(
-            reverse("accounts:signup"),
-            data={
-                "username": "test_user",
-                "email": "test_user@email.com",
-                "first_name": "test",
-                "last_name": "user",
-                "password1": "BDbdKDwpSt",
-                "password2": "BDbdKDwpStX",
-            },
-        )
-        self.assertContains(response, "Unsuccessful registration. Invalid information.")
-
-    def test_invalid_verification_link(self):
-        """
-        An invalid verification request should yield a redirect
-        """
-        response = self.client.get(reverse("accounts:activate", args=(1, 1)))
-        self.assertEqual(response.status_code, 302)
 
     def test_get_rating_one_restroom(self):
         """
@@ -254,29 +203,85 @@ class ViewTests(TestCase):
         self.assertEqual(len(Rating.objects.all()), 1)
         self.assertEqual(Rating.objects.all()[0].headline, "headline1")
 
-    # def test_rating_previously_rated_restroom(self):
-    #     """
-    #     Once a restroom has been rated, the same user should not be able
-    #     to rate it again
-    #     """
-    #     desc = "TEST DESCRIPTION"
-    #     yelp_id = "E6h-sMLmF86cuituw5zYxw"
-    #     rr = create_restroom(yelp_id, desc)
-    #     user = User.objects.create_user("Jon", "jon@email.com")
-    #     self.client.force_login(user=user)
-    #     Rating.objects.create(
-    #         restroom_id=rr,
-    #         user_id=user,
-    #         rating="4",
-    #         headline="headline1",
-    #         comment="comment1",
-    #     )
-    #     response = self.client.get(reverse("naturescall:rate_restroom", args=(1,)))
-    #     messages = [m.message for m in get_messages(response.wsgi_request)]
-    #     self.assertEqual(response.status_code, 302)
-    #     self.assertEqual(len(Rating.objects.all()), 1)
-    #     self.assertEqual(Rating.objects.all()[0].headline, "headline1")
-    #     self.assertIn(messages[0], "Sorry, You have already rated this restroom!!")
+    def test_seeing_previously_rated_restroom(self):
+        """
+        Once a restroom has been rated, the same user should be able
+        to see that rating
+        """
+        desc = "TEST DESCRIPTION"
+        yelp_id = "E6h-sMLmF86cuituw5zYxw"
+        rr = create_restroom(yelp_id, desc)
+        user = User.objects.create_user("Jon", "jon@email.com")
+        self.client.force_login(user=user)
+        Rating.objects.create(
+            restroom_id=rr,
+            user_id=user,
+            rating="4",
+            headline="headline1",
+            comment="comment1",
+        )
+        response = self.client.get(reverse("naturescall:rate_restroom", args=(1,)))
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(Rating.objects.all()[0].headline, "headline1")
+
+    def test_rating_previously_rated_restroom(self):
+        """
+        Once a restroom has been rated, that rating should be
+        editable using the restroom_detail link
+        """
+        desc = "TEST DESCRIPTION"
+        yelp_id = "E6h-sMLmF86cuituw5zYxw"
+        rr = create_restroom(yelp_id, desc)
+        user = User.objects.create_user("Jon", "jon@email.com")
+        self.client.force_login(user=user)
+        Rating.objects.create(
+            restroom_id=rr,
+            user_id=user,
+            rating="4",
+            headline="headline1",
+            comment="comment1",
+        )
+        response = self.client.post(
+            reverse("naturescall:rate_restroom", args=(1,)),
+            data={
+                "rating": "2",
+                "headline": "headline2",
+                "comment": "comment2",
+            },
+        )
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(len(Rating.objects.all()), 1)
+        self.assertEqual(Rating.objects.all()[0].headline, "headline2")
+
+    def test_delete_non_existent_rating(self):
+        """
+        user A will receive 404 message when trying to delete rating does not exist
+        """
+        user1 = User.objects.create_user("Simon1", "simon1@email.com")
+        self.client.force_login(user=user1)
+        response = self.client.get(reverse("naturescall:delete_rating", args=(1,)))
+        self.assertEqual(response.status_code, 404)
+
+    def test_delete_rating(self):
+        """
+        A user can delete rating
+        """
+        desc = "TEST DESCRIPTION"
+        yelp_id = "E6h-sMLmF86cuituw5zYxw"
+        rr = create_restroom(yelp_id, desc)
+        user = User.objects.create_user("Jon", "jon@email.com")
+        self.client.force_login(user=user)
+        Rating.objects.create(
+            restroom_id=rr,
+            user_id=user,
+            rating="4",
+            headline="headline1",
+            comment="comment1",
+        )
+        response = self.client.get(reverse("naturescall:delete_rating", args=(1,)))
+        self.assertEqual(response.status_code, 302)
+        messages = [m.message for m in get_messages(response.wsgi_request)]
+        self.assertIn(messages[0], "Your rating has been deleted!")
 
     def test_restroom_rating_calculation(self):
         """
